requirements.txt
python-dotenv
openai
requests
python-socketio[asyncio]
twitchio
elevenlabs
fastapi
uvicorn 